Added path saving to json Added Action Logs to display the past 10 actions

+++ Form1.cs
@@ -1,7 +1,13 @@
+using EnshroudedBackupManager.Models;
+using Newtonsoft.Json;
+
 namespace EnshroudedBackupManager;
 
 public partial class frmEnshroudedBackupManager : Form
 {
+    private List<string> _actionLogs = new List<string>(10);
+
+
     public frmEnshroudedBackupManager()
     {
         InitializeComponent();
@@ -9,23 +15,53 @@ public partial class frmEnshroudedBackupManager : Form
 
     private void frmEnshroudedBackupManager_Load(object sender, EventArgs e)
     {
-        var usersPath = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
+        string charFile, charBackupDir, worldFiles, worldBackupDir;
 
-        var charFile = Path.Combine(usersPath, "Saved Games", "Enshrouded", "characters");
-        var charBackupDir = Path.Combine(usersPath, "Saved Games", "Enshrouded", "Backup_Chars");
+        if (File.Exists(Path.Combine("Data", "filePaths.json")))
+        {
+            var filePathsJson = File.ReadAllText(Path.Combine("Data", "filePaths.json"));
+            var filePaths = JsonConvert.DeserializeObject<FilePaths>(filePathsJson);
 
-        var worldFilePaths = new List<string>();
-        worldFilePaths.Add(Path.Combine(usersPath, "Saved Games", "Enshrouded", "3ad85aea"));
-        worldFilePaths.Add(Path.Combine(usersPath, "Saved Games", "Enshrouded", "3ad85aea_info"));
+            if (filePaths == null)
+            {
+                setDefaultPaths();
+                return;
+            }
 
-        var worldFiles = string.Join(",", worldFilePaths);
-        var worldBackupDir = Path.Combine(usersPath, "Saved Games", "Enshrouded", "Backup_World");
+            txtCharFile.Text = filePaths.CharacterFilePath;
+            txtCharBackupDir.Text = filePaths.CharacterBackupDirectory;
+            txtWorldFiles.Text = filePaths.WorldFilePath;
+            txtWorldBackupDir.Text = filePaths.WorldBackupDirectory;
 
+            WriteActionLog("Paths Loaded from filePaths.json");
+        }
+        else
+        {
+            setDefaultPaths();
+            WritePathsToJson();
+        }
 
-        txtCharFile.Text = charFile;
-        txtCharBackupDir.Text = charBackupDir;
-        txtWorldFiles.Text = worldFiles;
-        txtWorldBackupDir.Text = worldBackupDir;
+        void setDefaultPaths()
+        {
+            var usersPath = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
+
+            charFile = Path.Combine(usersPath, "Saved Games", "Enshrouded", "characters");
+            charBackupDir = Path.Combine(usersPath, "Saved Games", "Enshrouded", "Backup_Chars");
+
+            var worldFilePaths = new List<string>();
+            worldFilePaths.Add(Path.Combine(usersPath, "Saved Games", "Enshrouded", "3ad85aea"));
+            worldFilePaths.Add(Path.Combine(usersPath, "Saved Games", "Enshrouded", "3ad85aea_info"));
+
+            worldFiles = string.Join(",", worldFilePaths);
+            worldBackupDir = Path.Combine(usersPath, "Saved Games", "Enshrouded", "Backup_World");
+
+            txtCharFile.Text = charFile;
+            txtCharBackupDir.Text = charBackupDir;
+            txtWorldFiles.Text = worldFiles;
+            txtWorldBackupDir.Text = worldBackupDir;
+
+            WriteActionLog("Default paths Loaded");
+        }
     }
 
     private void btnCharFile_Click(object sender, EventArgs e)
@@ -35,6 +71,8 @@ public partial class frmEnshroudedBackupManager : Form
         {
             var file = ofdSelectCharFileDialog.FileName;
             txtCharFile.Text = file;
+
+            WriteActionLog($"Character File Selected");
         }
     }
 
@@ -45,6 +83,8 @@ public partial class frmEnshroudedBackupManager : Form
         {
             var dir = fbdCharacterBackupDir.SelectedPath;
             txtCharBackupDir.Text = dir;
+
+            WriteActionLog("Character Backup Directory Selected");
         }
     }
 
@@ -55,6 +95,8 @@ public partial class frmEnshroudedBackupManager : Form
         {
             var files = ofdSelectWorldFiles.FileNames;
             txtWorldFiles.Text = string.Join(",", files);
+
+            WriteActionLog("World Files Selected");
         }
     }
 
@@ -65,6 +107,8 @@ public partial class frmEnshroudedBackupManager : Form
         {
             var dir = fbdWorldBackupDir.SelectedPath;
             txtWorldBackupDir.Text = dir;
+
+            WriteActionLog("World Backup Directory Selected");
         }
     }
 
@@ -75,15 +119,19 @@ public partial class frmEnshroudedBackupManager : Form
 
         if (string.IsNullOrWhiteSpace(charFile))
         {
+            WriteActionLog("ERROR: No Character File Selected");
             return;
         }
 
         if (string.IsNullOrWhiteSpace(charBackupDir))
         {
+            WriteActionLog("ERROR: No Character Backup Directory Selected");
             return;
         }
 
         File.Copy(charFile, Path.Combine(charBackupDir, Path.GetFileName(charFile)), true);
+
+        WriteActionLog("Character File Backed Up");
     }
 
     private void btnRestoreCharacter_Click(object sender, EventArgs e)
@@ -93,16 +141,20 @@ public partial class frmEnshroudedBackupManager : Form
 
         if (string.IsNullOrWhiteSpace(charFile))
         {
+            WriteActionLog("ERROR: No Character File Selected");
             return;
         }
 
         if (string.IsNullOrWhiteSpace(charBackupDir))
         {
+            WriteActionLog("ERROR: No Character Backup Directory Selected");
             return;
         }
 
         charBackupDir = Path.Combine(charBackupDir, "characters");
         File.Copy(charBackupDir, charFile, true);
+
+        WriteActionLog("Character File Restored");
     }
 
     private void btnBackupWorld_Click(object sender, EventArgs e)
@@ -112,19 +164,35 @@ public partial class frmEnshroudedBackupManager : Form
 
         if (worldFiles.Length <= 0)
         {
+            WriteActionLog("ERROR: No World Files Selected");
             return;
         }
 
         if (string.IsNullOrWhiteSpace(worldBackupDir))
         {
+            WriteActionLog("ERROR: No World Backup Directory Selected");
             return;
+        }
+
+        if (!Directory.Exists(worldBackupDir))
+        {
+            Directory.CreateDirectory(worldBackupDir);
+            WriteActionLog("World Backup Directory Created");
         }
 
         foreach (var file in worldFiles)
         {
-            File.Copy(file, Path.Combine(worldBackupDir, Path.GetFileName(file)), true);
-
+            if (File.Exists(file))
+            {
+                File.Copy(file, Path.Combine(worldBackupDir, Path.GetFileName(file)), true);
+            }
+            else
+            {
+                WriteActionLog("ERROR: World File Not Found");
+            }
         }
+
+        WriteActionLog("World Files Backed Up");
     }
 
     private void btnRestoreWorld_Click(object sender, EventArgs e)
@@ -134,17 +202,75 @@ public partial class frmEnshroudedBackupManager : Form
 
         if (worldFiles.Length <= 0)
         {
+            WriteActionLog("ERROR: No World Files Selected");
             return;
         }
 
         if (string.IsNullOrWhiteSpace(worldBackupDir))
         {
+            WriteActionLog("ERROR: No World Backup Directory Selected");
             return;
         }
 
         foreach (var file in worldFiles)
         {
-            File.Copy(Path.Combine(worldBackupDir, Path.GetFileName(file)), file, true);
+            if (File.Exists(file) && Directory.Exists(worldBackupDir))
+            {
+                File.Copy(Path.Combine(worldBackupDir, Path.GetFileName(file)), file, true);
+            }
+            else
+            {
+                WriteActionLog("ERROR: Enshrouded or Backup Directory Not Found");
+                return;
+            }
         }
+
+        WriteActionLog("World Files Restored");
+    }
+
+    private void WriteActionLog(string action)
+    {
+        var actionLog = $"{DateTime.Now.ToString("HH:mm:ss")}: {action}";
+        _actionLogs.Add(actionLog);
+
+        // if more than 10 logs are stored in _actionLogs, remove the oldest one
+        if (_actionLogs.Count() > 10)
+        {
+            _actionLogs = _actionLogs.Skip(1).ToList();
+        }
+
+        // copy the _actionLogs list to a new list so that it can be reversed
+        var reversed_actionLogs = _actionLogs.ToList();
+        reversed_actionLogs.Reverse();
+
+        rtbActionLog.Text = string.Join(Environment.NewLine, reversed_actionLogs);
+    }
+
+    private void WritePathsToJson()
+    {
+        // create the Data directory if it doesn't exist
+        if (!Directory.Exists("./Data"))
+        {
+            Directory.CreateDirectory("./Data");
+            WriteActionLog("Data Directory Created");
+        }
+
+        var filePaths = new FilePaths
+        {
+            CharacterFilePath = txtCharFile.Text,
+            CharacterBackupDirectory = txtCharBackupDir.Text,
+            WorldFilePath = txtWorldFiles.Text,
+            WorldBackupDirectory = txtWorldBackupDir.Text
+        };
+
+        var filePathsJson = JsonConvert.SerializeObject(filePaths, Formatting.Indented);
+        File.WriteAllText(Path.Combine("Data", "filePaths.json"), filePathsJson);
+
+        WriteActionLog("Paths written to filePaths.json");
+    }
+
+    private void txtCharFile_Leave(object sender, EventArgs e)
+    {
+        WritePathsToJson();
     }
 }
